requirements.txt
tensorflow-gpu==1.9.0
# Keras>=2.2.2
# tensorboard>=1.10.0
h5py==2.8.0
matplotlib==2.2.2
numpy<=1.14.5
scikit-learn==0.20.0
wget>=3.2
tqdm>=4.23.2
sty>=1.0.0b6
requests>=2.19
opencv-python>=3.4.0.12
image>=1.5.20
pandas>=0.23.4
scikit-image>=0.14.0
dropbox>=9.0.0
imgaug>=0.2.6
imageio>=2.4.1